implement basic check flag
This implements a very basic version of a check flag. If any
dependencies that should move to the workspace are found, they are
printed to stderr and the command exits with a FAILURE `ExitCode`.

If any dependencies are found in a workspace member, but that have an
existing workspace dependency that can be directly used (that is if the
workspace Cargo.toml doesn't need to be modified) nothing is printed,
but the command still exits with FAILURE.

Ideally the check flag would work like `cargo fmt --check` where a diff
is shown, but that would require extensive refactoring.