Add flag for dry-run
With the dry-run capability, cargo-autoinherit can be run in CI and used
to enforce strict inheritance except for excluded workspace members.

The dry run will not touch any files in the workspace. However if any
file would be touched or there are any dependencies which could not be
autoinherited, it returns with a non-zero exit code.